add all required data plus implementing feign client

+++ customer-service/src/main/java/org/korcula/repository/CustomerRepository.java
@@ -3,6 +3,7 @@ package org.korcula.repository;
 import org.korcula.model.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
 
+//	List<ResponseCustomerAndProductName> getCustomerAndProductName();
 }
